put large real data sample in separate file
